Info about Win32 structure, Glob exp. on POSIX
Added information about how the Win32 structure works and how is it built and more information about linking and stuff.
DLLs are pre-built *for now* because it's a lot of *fuss* building and linking them properly. If wish to omit the pre-packed DLLs
use 'vcpkg' or built them yourselves.

Glob expressions are only supported in the UNIX version of the program. If you wish to build this with Glob expressions
either stick to the UNIX way with MinGW/MSYS/Cygwin or implement a way for the Win32 yourself.